+++ matchbox-server/src/main/java/ch/ahdis/matchbox/validation/ValidationProvider.java
@@ -94,10 +94,6 @@ public class ValidationProvider {
 
 	@Autowired
 	private MatchboxImplementationGuideProvider igProvider;
-
-	@Autowired
-	private INpmPackageVersionDao myPackageVersionDao;
-
 	@Autowired
 	private INpmPackageVersionResourceDao myPackageVersionResourceDao;