+++ docs/changelog.md
@@ -3,6 +3,9 @@ Unreleased
 - Allow providing map and models in the StructureMap $transform operation [#305](https://github.com/ahdis/matchbox/issues/305)
 - Introduce parameter 'autoInstallMissingIgs' to automatically install IGs from the public registry
   [#306](https://github.com/ahdis/matchbox/issues/306)
+- Introduce the configuration parameter 'devMode' to enable the development environment; it allows installing an 
+  ImplementationGuide by posting its NPM package to the operation _$install-npm-package_
+  [#306](https://github.com/ahdis/matchbox/issues/306)
 
 2024/11/13 Release 3.9.7